requirements.txt
django
django-cors-headers
python-decouple
djangorestframework
djangorestframework-simplejwt
requests
pillow
python-dateutil